refactored section-box component and finished infographics-view component
