+++ frontend/src/components/section-box/section-box-component.js
