+++ frontend/src/components/section-box/section-box.module.scss
